derivative utility function (needs doubles)

+++ bullethellwhatever/UtilitySystems/Utilities.cs
@@ -191,11 +191,11 @@ namespace bullethellwhatever
             return (float)generated;
         }
 
-        public static float DerivativeOfFunctionAtTime(Func<float, float> function, float time)
+        public static float DerivativeOfFunctionAtTime(Func<double, double> function, float time)
         {
-            float h = 1e-6f;
+            double h = 1e-6f;
 
-            return (function(time + h) - function(time)) / h;
+            return (float)((function((double)time + h) - function((double)time)) / h);
         }
         public static bool RandomChance(int oneIn)
         {
